p30-1 フェーズごとにファイルを分割、ファイルごとにテスト作成 : [test]chapter1/statement,createStatementData > green
