+++ src/chapter1/statement.js
@@ -1,0 +1,28 @@
+import { createStatementData } from "./createStatementData.js"
+
+export function statement(invoice, plays) {
+    return renderPlainText(createStatementData(invoice, plays));
+}
+
+export function renderPlainText(data) {
+  let result = `Statement for ${data.customer}\n`;
+
+  for (let perf of data.performances) {
+    // 注文の内訳を出力
+    result += ` ${perf.play.name}: ${usd(perf.amount)} (${
+      perf.audience
+    } seats)\n`;
+  }
+
+  result += `Amount owed is ${usd(data.totalAmount)}\n`;
+  result += `You earned ${data.totalVolumeCredits} credits\n`;
+  return result;
+
+  function usd(aNumber) {
+    return new Intl.NumberFormat("en-US", {
+      style: "currency",
+      currency: "USD",
+      minimumFractionDigits: 2,
+    }).format(aNumber / 100);
+  }
+}
